refactor(patcher): carry code addresses as uintptr_t
Every code address in the engine was a uint32_t, which held while both targets
were 32-bit but stops the core compiling for the macOS x86_64 dylib, where a
hook address is a virtual address above 4 GB. Addresses become uintptr_t, and
the config parser now reads a 64-bit value and refuses one that does not fit
the target's pointer instead of truncating it into a plausible-looking address
elsewhere in the image.

- include/patcher.h, include/wrappers/wrapper_base.h: hookAddress and
  consumedExitAddress widened.
- include/trampoline.h, src/core/trampoline.cpp: address parameters widened;
  the rel32 deltas are computed through uintptr_t and narrowed explicitly.
- src/core/config_reader.cpp: ParseHexAddress becomes ParseHexValue over
  uint64_t with an ERANGE check, and NarrowCodeAddress gates the store.
- src/core/patcher.cpp: address diagnostics use PRIXPTR so one format string
  is correct on both widths.

+++ src/KotorPatcher/include/trampoline.h
@@ -17,20 +17,23 @@
 //
 // The unprotect/write/reprotect and instruction-cache handling live behind
 // Platform::WriteCode, so this module is pure x86 byte layout.
+//
+// Addresses are uintptr_t because the engine dereferences them: 32 bits on the
+// i686/i386 targets, 64 on the macOS x86_64 dylib.
 
 namespace KotorPatcher {
     namespace Trampoline {
         // Write a 5-byte relative JMP at `address` targeting `target`.
-        bool WriteJump(uint32_t address, void* target);
+        bool WriteJump(uintptr_t address, void* target);
 
         // Write a 5-byte relative CALL at `address` targeting `target`.
-        bool WriteCall(uint32_t address, void* target);
+        bool WriteCall(uintptr_t address, void* target);
 
         // Compare `length` bytes at `address` against `expected`. Used as the
         // wrong-game-version guard before any patch is applied.
-        bool VerifyBytes(uint32_t address, const uint8_t* expected, std::size_t length);
+        bool VerifyBytes(uintptr_t address, const uint8_t* expected, std::size_t length);
 
         // Overwrite `length` bytes at `startAddress` with NOPs (0x90).
-        bool WriteNoOps(uint32_t startAddress, std::size_t length);
+        bool WriteNoOps(uintptr_t startAddress, std::size_t length);
     }
 }
